requirements.txt
streamlit==1.32.0
openai==1.12.0
python-dotenv==1.0.1
pillow==10.2.0
sqlite3==2.6.0
hashlib==3.2
